requirements.txt
flask-restx==0.2.0
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.4
flake8==3.8.4
psycopg2-binary==2.8.6
pandas==1.1.5
validator==0.7.0